created not found excepcion

+++ src/main/java/com/jesus/sales/controller/CategoryController.java
@@ -1,5 +1,6 @@
 package com.jesus.sales.controller;
 
+import com.jesus.sales.exception.ModelNotFoundException;
 import com.jesus.sales.model.Category;
 import com.jesus.sales.service.ICategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +27,9 @@ public class CategoryController {
     @GetMapping("/{id}")
     public ResponseEntity<Category> readById(@PathVariable("id") Integer id) throws  Exception{
         Category obj = service.readById(id);
+        if (obj == null){
+            throw new ModelNotFoundException("ID NOT FOUND" + id);
+        }
         return new ResponseEntity<>(obj, HttpStatus.OK);
     }
 
@@ -43,6 +47,10 @@ public class CategoryController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable("id") Integer id) throws  Exception{
+        Category obj = service.readById(id);
+        if (obj == null){
+            throw new ModelNotFoundException("ID NOT FOUND" + id);
+        }
         service.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
